# Add a faster string-search method with simd instructions than std::find