### Description
In practice, c_strstr is always faster than std::find in most cases, refer to [1]. However, the c_strstr interface will truncate '\0', which makes it unable to handle unicode characters. In addition, glibc uses two-way- search, see[2], but it does not seem to be exposed as a common interface. It requires some additional code to determine the best path. In some scenarios, the decision-making code may cause performance degradation.

### Solution
There are many optimization algorithms for string-search, refer to [7], but they cannot change the essence of the problem, which is a search process of O(n*m) complexity. I think the most common way to improve performance is to optimize the search instructions. For example, doris (the same is true for clickhouse, this part of the code of doris was copied from clickhouse, and the Volnitsky[11] part was removed), see [5].
This PR refers to the implementation of [5], [6], [8] ：
[5] doris : first-two comapre first, and  the issue of page-safe was also mentioned
[6] stringzilla: first-mid-lst compare first
[8] simd-strstr implement by WojciechMula, first-lst comapre first.

They all look few chars first, the more chars you choose to look the higher probability of hitting the optimized path, and the more cost  of 'optimized path' itself.

In the end, the solution I chose is to first compare first-last-chars and then compare the remaining bytes. This is because in practice, our users seem to be able to get the matching results through first-last-chars.

**ut & benchmark**
This part of the code refers to folly’s test code, see [9] [10], of course, some adjustments have been made based on the code implementation.

### Benchmark : std::find vs simdStrStr:

findSuccessful(opt_50_5)                                   31.66ms     31.59
findSuccessful(opt_100_10)                                 36.81ms     27.16
findSuccessful(opt_100_20)                                 96.47ms     10.37
findSuccessful(opt_1k_10)                                  63.27ms     15.81
findSuccessful(opt_1k_100)                                156.95ms      6.37
findSuccessful(std_50_5)                                   45.24ms     22.11
findSuccessful(std_100_10)                                191.23ms      5.23
findSuccessful(std_100_20)                                322.11ms      3.10
findSuccessful(std_1k_10)                                 122.69ms      8.15
findSuccessful(std_1k_100)                                168.10ms      5.95
findUnsuccessful(std_first_char_match)                    819.13ms      1.22
findUnsuccessful(opt_first_char_match)                    261.46ms      3.82
findUnsuccessful(std_first_char_unmatch)                  249.12ms      4.01
findUnsuccessful(opt_first_char_unmatch)                  199.02ms      5.02

### References

[1] gcc std::search issue:
https://gcc.gnu.org/bugzilla/show_bug.cgi?id=66414
[2] glibc strstr implements:
https://github.com/lattera/glibc/blob/master/string/strstr.c
[3] two way search:
http://www-igm.univ-mlv.fr/~lecroq/string/node26.html#SECTION00260
[4] std::boyer_moore_searcher
https://en.cppreference.com/w/cpp/utility/functional/boyer_moore_searcher
https://github.com/mclow/search-library?tab=readme-ov-file
[5] doris string-search
https://github.com/apache/doris/blob/4e326e5f496109edb253d094869d057573ab7a59/be/src/vec/common/string_searcher.h#L44
[6] stringzilla
https://github.com/ashvardanian/StringZilla/blob/main/include/stringzilla/stringzilla.h#L2187
[7] string-search algorithms
https://www-igm.univ-mlv.fr/~lecroq/string/
[8] sse4 strst
https://github.com/WojciechMula/sse4-strstr
[9] folly benchmark
https://github.com/facebook/folly/blob/ce5edfb9b08ead9e78cb46879e7b9499861f7cd2/folly/test/FBStringTestBenchmarks.cpp.h
[10] folly ut
https://github.com/facebook/folly/blob/ce5edfb9b08ead9e78cb46879e7b9499861f7cd2/folly/test/FBStringTest.cpp#L1277
[11] clickhouse Volnitsky
https://github.com/ClickHouse/ClickHouse/blob/a51f867ce014a4cb8f97c46e004b90f5feafd80f/src/Common/Volnitsky.h#L482