@skadilover 2024-08-29
@Yuhta 
ping~

> I don't think you can put variable definition in header file, you probably need to move it to SimdUtil.cpp and have only forward declaration here. Also what is the value if we do MADV_HUGEPAGE for some addresses?

I found that the performance of the code will be reduced by 30% when placed in the cpp file. Could you give me some suggestions to take into account these two issues?