@skadilover 2024-09-09
@Yuhta  
> Please consider the other suggestions in https://github.com/facebookincubator/velox/pull/10731#pullrequestreview-2255002180

Updates code for the suggestions in #10731:

> These checks can be VELOX_DCHECK since their correctness can be ensured quite locally in the same file


done

> VELOX_SIMD_STRSTR_CASE


done

> aside: Wondering what is the minimal k to make KMP faster than this. This is also related to content of the strings though.

I  have extended StringBenchmark to support different algrithm, and kmp is always slower than simdStrStr:

```
findSuccessful(simd_50_to_10)                             841.16ms      1.19
findSuccessful(std_50_to_10)                              518.25ms      1.93
findSuccessful(kmp_50_to_10)                                 3.47s   288.43m
findSuccessful(simd_100_to_10)                            691.35ms      1.45
findSuccessful(std_100_to_10)                             517.59ms      1.93
findSuccessful(kmp_100_to_10)                                2.49s   401.61m
findSuccessful(simd_100_to_20)                               1.38s   723.47m
findSuccessful(std_100_to_20)                             840.38ms      1.19
findSuccessful(kmp_100_to_20)                                5.27s   189.58m
findSuccessful(simd_1000_to_10)                             11.93s    83.82m
findSuccessful(std_1000_to_10)                               4.48s   223.43m
findSuccessful(kmp_1000_to_10)                             1.03min    16.14m
findSuccessful(simd_1000_to_100)                            13.01s    76.87m
findSuccessful(std_1000_to_100)                             14.18s    70.54m
findSuccessful(kmp_1000_to_100)                            1.94min     8.61m
findSuccessful(simd_1000_to_200)                             7.13s   140.25m
findSuccessful(std_1000_to_200)                             10.80s    92.60m
findSuccessful(kmp_1000_to_200)                             32.50s    30.77m
findUnsuccessful(std_first_char_match)                    996.51ms      1.00
findUnsuccessful(opt_first_char_match)                    330.99ms      3.02
findUnsuccessful(std_first_char_unmatch)                  415.35ms      2.41
findUnsuccessful(opt_first_char_unmatch)                  275.87ms      3.62
```


> Since this is an almost drop-in replacement for strstr, can you add another test to generate a few thousands of random string pairs with various lengths, and check our result is the same as strstr? Also make sure you cover corner cases like larger needle than haystack, etc.

I update ut refer to folly uts, add random test cases.

Could you have look at this once more?