@skadilover 2024-09-13
@Yuhta  update code 
1. pageSize => global variable
2. fix comments above except this:

>Will we miss the opportunity if it is a super long string that is spanning multiple pages?

As we inline kPageSize,  pageSafe method is faster than before, I move the check to each batch iteration, we only lose the comparison-span-pages case, maybe better,  like clickhouse:
https://github.com/ClickHouse/ClickHouse/blob/master/src/Common/StringSearcher.h

```
            while (haystack < haystack_end)
            {
                if (haystack + N <= haystack_end && isPageSafe(haystack))
                {
```


new benchmark result: 

```
findSuccessful(simd_50_to_5)                              355.40ms      2.81
findSuccessful(std_50_to_5)                               807.73ms      1.24
findSuccessful(std_boyer_moore_50_to_5)                      1.40s   712.05m
findSuccessful(kmp_50_to_5)                                  5.18s   192.87m
findSuccessful(simd_100_to_10)                            319.76ms      3.13
findSuccessful(std_100_to_10)                                1.25s   802.85m
findSuccessful(std_boyer_moore_100_to_10)                 910.54ms      1.10
findSuccessful(kmp_100_to_10)                                6.38s   156.72m
findSuccessful(simd_100_to_20)                            355.43ms      2.81
findSuccessful(std_100_to_20)                                2.07s   483.83m
findSuccessful(std_boyer_moore_100_to_20)                 901.23ms      1.11
findSuccessful(kmp_100_to_20)                                6.20s   161.18m
findSuccessful(simd_1000_to_10)                           245.61ms      4.07
findSuccessful(std_1000_to_10)                            724.69ms      1.38
findSuccessful(std_boyer_moore_1000_to_10)                870.93ms      1.15
findSuccessful(kmp_1000_to_10)                               3.33s   300.34m
findSuccessful(simd_1000_to_100)                           69.67ms     14.35
findSuccessful(std_1000_to_100)                           186.06ms      5.37
findSuccessful(std_boyer_moore_1000_to_100)               286.45ms      3.49
findSuccessful(kmp_1000_to_100)                           913.97ms      1.09
findSuccessful(simd_1000_to_200)                          354.15ms      2.82
findSuccessful(std_1000_to_200)                           642.50ms      1.56
findSuccessful(std_boyer_moore_1000_to_200)               675.70ms      1.48
findSuccessful(kmp_1000_to_200)                              9.78s   102.28m
findUnsuccessful(std_first_char_match)                    920.16ms      1.09
findUnsuccessful(opt_first_char_match)                    383.93ms      2.60
findUnsuccessful(std_first_char_unmatch)                  334.28ms      2.99
findUnsuccessful(opt_first_char_unmatch)                  331.98ms      3.01

```